Resolve #98, #99
Resolve #98 Refactor reducers
Resolve #99 Refactor actions
