+++ src/actions/userActions.js
@@ -1,0 +1,13 @@
+function onLogin (loginData) {
+  return {
+    type: 'LOGIN',
+    payload: {
+      profile: {
+        ...loginData,
+        isAuthenticated: true
+      }
+    }
+  }
+}
+
+export { onLogin }
