+++ src/reducers/transferReducer.js
@@ -1,46 +1,43 @@
-import update from 'immutability-helper'
+/*
+ *  Handle transfer actions amd transfer data
+ *
+ *  e.g. tx hash, timestamp, sender, receiver, etc
+ */
 
 const initialState = {
-  walletSelection: null,
-  cryptoSelection: null,
-  transferForm: {
-    transferAmount: '',
-    password: '',
-    destination: '',
-    sender: ''
-  }
+  // transfer data (sender, destination, encryptedWallet, etc)
+  // fetched from database
+  transfer: null,
+
+  // gas related data
+  gasCost: null,
+
+  // transaction receipt
+  receipt: null
 }
 
 export default function (state = initialState, action) {
   switch (action.type) {
-    case 'SELECT_CRYPTO':
-      return {
-        ...state,
-        cryptoSelection: state.cryptoSelection === action.payload ? null : action.payload
-      }
-    case 'SELECT_WALLET':
-      return {
-        ...state,
-        walletSelection: state.walletSelection === action.payload ? null : action.payload,
-        cryptoSelection: null
-      }
-    case 'UPDATE_TRANSFER_FORM':
-      return {
-        ...state,
-        transferForm: action.payload
-      }
-    case 'GENERATE_SECURITY_ANSWER':
-      return update(state, { transferForm: { password: { $set: action.payload } } })
     case 'GET_TRANSFER_FULFILLED':
       return {
         ...state,
         transfer: action.payload
       }
-    case 'VERIFY_PASSWORD_FULFILLED':
-    // store decrypted wallet in transfer.wallet
-      return update(state, { transfer: { wallet: { $set: action.payload } } })
-    case 'CLEAR_DECRYPTED_WALLET':
-      return update(state, { transfer: { wallet: { $set: null } } })
+    case 'GET_GAS_COST_FULFILLED':
+      return {
+        ...state,
+        gasCost: action.payload
+      }
+    case 'TRANSACTION_HASH_RETRIEVED_FULFILLED':
+      return {
+        ...state,
+        receipt: action.payload
+      }
+    case 'ACCEPT_TRANSFER_TRANSACTION_HASH_RETRIEVED_FULFILLED':
+      return {
+        ...state,
+        receipt: action.payload
+      }
     default: // need this for default case
       return state
   }
